mv dashboard page tour fixture to browser tests
